+++ base/src/main/java/proguard/classfile/editor/ClassReferenceFixer.java
@@ -84,6 +84,7 @@ import proguard.classfile.kotlin.KotlinAnnotationArgument;
 import proguard.classfile.kotlin.KotlinClassKindMetadata;
 import proguard.classfile.kotlin.KotlinConstructorMetadata;
 import proguard.classfile.kotlin.KotlinDeclarationContainerMetadata;
+import proguard.classfile.kotlin.KotlinEnumEntryMetadata;
 import proguard.classfile.kotlin.KotlinFileFacadeKindMetadata;
 import proguard.classfile.kotlin.KotlinFunctionMetadata;
 import proguard.classfile.kotlin.KotlinMetadata;
@@ -95,13 +96,15 @@ import proguard.classfile.kotlin.KotlinTypeAliasMetadata;
 import proguard.classfile.kotlin.KotlinTypeMetadata;
 import proguard.classfile.kotlin.KotlinTypeParameterMetadata;
 import proguard.classfile.kotlin.KotlinValueParameterMetadata;
-import proguard.classfile.kotlin.flags.KotlinPropertyAccessorFlags;
+import proguard.classfile.kotlin.flags.KotlinPropertyAccessorMetadata;
 import proguard.classfile.kotlin.visitor.AllTypeVisitor;
 import proguard.classfile.kotlin.visitor.KotlinAnnotationArgumentVisitor;
 import proguard.classfile.kotlin.visitor.KotlinAnnotationVisitor;
 import proguard.classfile.kotlin.visitor.KotlinConstructorVisitor;
+import proguard.classfile.kotlin.visitor.KotlinEnumEntryVisitor;
 import proguard.classfile.kotlin.visitor.KotlinFunctionVisitor;
 import proguard.classfile.kotlin.visitor.KotlinMetadataVisitor;
+import proguard.classfile.kotlin.visitor.KotlinPropertyAccessorVisitor;
 import proguard.classfile.kotlin.visitor.KotlinPropertyVisitor;
 import proguard.classfile.kotlin.visitor.KotlinTypeAliasVisitor;
 import proguard.classfile.kotlin.visitor.KotlinTypeParameterVisitor;
@@ -542,12 +545,14 @@ public class ClassReferenceFixer
           KotlinPropertyVisitor,
           KotlinFunctionVisitor,
           KotlinConstructorVisitor,
+          KotlinEnumEntryVisitor,
           KotlinTypeVisitor,
           KotlinTypeAliasVisitor,
           KotlinValueParameterVisitor,
           KotlinTypeParameterVisitor,
           KotlinAnnotationVisitor,
-          KotlinAnnotationArgumentVisitor {
+          KotlinAnnotationArgumentVisitor,
+          KotlinPropertyAccessorVisitor {
     // Implementations for KotlinMetadataVisitor.
     @Override
     public void visitAnyKotlinMetadata(Clazz clazz, KotlinMetadata kotlinMetadata) {}
@@ -586,10 +591,8 @@ public class ClassReferenceFixer
                 kotlinClassKindMetadata.referencedCompanionClass);
       }
 
-      for (int k = 0; k < kotlinClassKindMetadata.enumEntryNames.size(); k++) {
-        kotlinClassKindMetadata.enumEntryNames.set(
-            k, kotlinClassKindMetadata.referencedEnumEntries.get(k).getName(clazz));
-      }
+      kotlinClassKindMetadata.enumEntries.forEach(
+          enumEntry -> enumEntry.name = enumEntry.referencedEnumEntry.getName(clazz));
 
       for (int k = 0; k < kotlinClassKindMetadata.nestedClassNames.size(); k++) {
         kotlinClassKindMetadata.nestedClassNames.set(
@@ -613,6 +616,7 @@ public class ClassReferenceFixer
       kotlinClassKindMetadata.superTypesAccept(clazz, this);
       kotlinClassKindMetadata.constructorsAccept(clazz, this);
       kotlinClassKindMetadata.inlineClassUnderlyingPropertyTypeAccept(clazz, this);
+      kotlinClassKindMetadata.annotationsAccept(clazz, this);
 
       visitKotlinDeclarationContainerMetadata(clazz, kotlinClassKindMetadata);
     }
@@ -666,19 +670,21 @@ public class ClassReferenceFixer
             new FieldSignature(backingFieldClass, backingField);
       }
 
-      kotlinPropertyMetadata.getterSignature =
+      kotlinPropertyMetadata.getterMetadata.signature =
           fixPropertyMethod(
               programClass,
-              kotlinPropertyMetadata.referencedGetterMethod,
-              kotlinPropertyMetadata.getterFlags,
-              kotlinPropertyMetadata.getterSignature);
+              kotlinPropertyMetadata.getterMetadata.referencedMethod,
+              kotlinPropertyMetadata.getterMetadata,
+              kotlinPropertyMetadata.getterMetadata.signature);
 
-      kotlinPropertyMetadata.setterSignature =
-          fixPropertyMethod(
-              programClass,
-              kotlinPropertyMetadata.referencedSetterMethod,
-              kotlinPropertyMetadata.setterFlags,
-              kotlinPropertyMetadata.setterSignature);
+      if (kotlinPropertyMetadata.setterMetadata != null) {
+        kotlinPropertyMetadata.setterMetadata.signature =
+            fixPropertyMethod(
+                programClass,
+                kotlinPropertyMetadata.setterMetadata.referencedMethod,
+                kotlinPropertyMetadata.setterMetadata,
+                kotlinPropertyMetadata.setterMetadata.signature);
+      }
 
       kotlinPropertyMetadata.syntheticMethodForAnnotations =
           fixPropertyMethod(
@@ -699,8 +705,10 @@ public class ClassReferenceFixer
       kotlinPropertyMetadata.contextReceiverTypesAccept(
           clazz, kotlinDeclarationContainerMetadata, this);
       kotlinPropertyMetadata.typeAccept(clazz, kotlinDeclarationContainerMetadata, this);
-      kotlinPropertyMetadata.setterParametersAccept(
+      kotlinPropertyMetadata.setterParameterAccept(clazz, kotlinDeclarationContainerMetadata, this);
+      kotlinPropertyMetadata.propertyAccessorsAccept(
           clazz, kotlinDeclarationContainerMetadata, this);
+      kotlinPropertyMetadata.annotationsAccept(clazz, this);
     }
 
     // Implementations for KotlinFunctionVisitor.
@@ -725,6 +733,7 @@ public class ClassReferenceFixer
       kotlinFunctionMetadata.valueParametersAccept(clazz, kotlinMetadata, this);
       kotlinFunctionMetadata.returnTypeAccept(clazz, kotlinMetadata, this);
       kotlinFunctionMetadata.contractsAccept(clazz, kotlinMetadata, new AllTypeVisitor(this));
+      kotlinFunctionMetadata.annotationsAccept(clazz, this);
 
       String newFunctionName =
           newKotlinFunctionName(
@@ -749,12 +758,12 @@ public class ClassReferenceFixer
       }
 
       kotlinConstructorMetadata.valueParametersAccept(clazz, kotlinClassKindMetadata, this);
+      kotlinConstructorMetadata.annotationsAccept(clazz, this);
     }
 
     // Implementations for KotlinTypeVisitor.
     @Override
     public void visitAnyType(Clazz clazz, KotlinTypeMetadata kotlinTypeMetadata) {
-
       if (kotlinTypeMetadata.className != null) {
         kotlinTypeMetadata.className = kotlinTypeMetadata.referencedClass.getName();
       }
@@ -774,7 +783,6 @@ public class ClassReferenceFixer
         Clazz clazz,
         KotlinDeclarationContainerMetadata kotlinDeclarationContainerMetadata,
         KotlinTypeAliasMetadata kotlinTypeAliasMetadata) {
-
       kotlinTypeAliasMetadata.annotationsAccept(clazz, this);
       kotlinTypeAliasMetadata.underlyingTypeAccept(clazz, kotlinDeclarationContainerMetadata, this);
       kotlinTypeAliasMetadata.expandedTypeAccept(clazz, kotlinDeclarationContainerMetadata, this);
@@ -784,7 +792,9 @@ public class ClassReferenceFixer
     // Implementations for KotlinValueParameterVisitor
     @Override
     public void visitAnyValueParameter(
-        Clazz clazz, KotlinValueParameterMetadata kotlinValueParameterMetadata) {}
+        Clazz clazz, KotlinValueParameterMetadata kotlinValueParameterMetadata) {
+      kotlinValueParameterMetadata.annotationsAccept(clazz, this);
+    }
 
     @Override
     public void visitFunctionValParameter(
@@ -793,6 +803,7 @@ public class ClassReferenceFixer
         KotlinFunctionMetadata kotlinFunctionMetadata,
         KotlinValueParameterMetadata kotlinValueParameterMetadata) {
       kotlinValueParameterMetadata.typeAccept(clazz, kotlinMetadata, kotlinFunctionMetadata, this);
+      visitAnyValueParameter(clazz, kotlinValueParameterMetadata);
     }
 
     @Override
@@ -803,6 +814,7 @@ public class ClassReferenceFixer
         KotlinValueParameterMetadata kotlinValueParameterMetadata) {
       kotlinValueParameterMetadata.typeAccept(
           clazz, kotlinClassKindMetadata, kotlinConstructorMetadata, this);
+      visitAnyValueParameter(clazz, kotlinValueParameterMetadata);
     }
 
     @Override
@@ -813,6 +825,7 @@ public class ClassReferenceFixer
         KotlinValueParameterMetadata kotlinValueParameterMetadata) {
       kotlinValueParameterMetadata.typeAccept(
           clazz, kotlinDeclarationContainerMetadata, kotlinPropertyMetadata, this);
+      visitAnyValueParameter(clazz, kotlinValueParameterMetadata);
     }
 
     // Implementations for KotlinTypeParameterVisitor
@@ -821,6 +834,24 @@ public class ClassReferenceFixer
         Clazz clazz, KotlinTypeParameterMetadata kotlinTypeParameterMetadata) {
       kotlinTypeParameterMetadata.annotationsAccept(clazz, this);
       kotlinTypeParameterMetadata.upperBoundsAccept(clazz, this);
+      kotlinTypeParameterMetadata.annotationsAccept(clazz, this);
+    }
+
+    @Override
+    public void visitAnyEnumEntry(
+        Clazz clazz,
+        KotlinClassKindMetadata kotlinClassKindMetadata,
+        KotlinEnumEntryMetadata kotlinEnumEntryMetadata) {
+      kotlinEnumEntryMetadata.annotationsAccept(clazz, this);
+    }
+
+    @Override
+    public void visitAnyPropertyAccessor(
+        Clazz clazz,
+        KotlinMetadata kotlinMetadata,
+        KotlinPropertyMetadata kotlinPropertyMetadata,
+        KotlinPropertyAccessorMetadata kotlinPropertyAccessorMetadata) {
+      kotlinPropertyAccessorMetadata.annotationsAccept(clazz, this);
     }
 
     // Implementations for KotlinAnnotationVisitor
@@ -828,7 +859,9 @@ public class ClassReferenceFixer
     @Override
     public void visitAnyAnnotation(
         Clazz clazz, KotlinAnnotatable annotatable, KotlinAnnotation annotation) {
-      annotation.className = annotation.referencedAnnotationClass.getName();
+      if (annotation.referencedAnnotationClass != null) {
+        annotation.className = annotation.referencedAnnotationClass.getName();
+      }
 
       annotation.argumentsAccept(clazz, annotatable, this);
     }
@@ -1039,7 +1072,7 @@ public class ClassReferenceFixer
   private static MethodSignature fixPropertyMethod(
       Clazz referencedMethodClass,
       Method referencedMethod,
-      KotlinPropertyAccessorFlags flags,
+      KotlinPropertyAccessorMetadata accessorMetadata,
       MethodSignature oldSignature) {
     if (oldSignature == null) {
       return null;
@@ -1047,9 +1080,9 @@ public class ClassReferenceFixer
 
     MethodSignature newSignature = new MethodSignature(referencedMethodClass, referencedMethod);
 
-    if (!oldSignature.equals(newSignature) && flags != null) {
+    if (!oldSignature.equals(newSignature) && accessorMetadata != null) {
       addJvmNameAnnotation((ProgramClass) referencedMethodClass, (ProgramMethod) referencedMethod);
-      flags.hasAnnotations = true;
+      accessorMetadata.hasAnnotations = true;
     }
 
     return newSignature;