+++ base/src/main/java/proguard/classfile/kotlin/reflect/PropertyReferenceInfo.java
@@ -61,8 +61,9 @@ public class PropertyReferenceInfo implements CallableReferenceInfo {
    */
   @Override
   public String getSignature() {
-    return propertyMetadata.getterSignature != null
-        ? propertyMetadata.getterSignature.method + propertyMetadata.getterSignature.descriptor
+    return propertyMetadata.getterMetadata.signature != null
+        ? propertyMetadata.getterMetadata.signature.method
+            + propertyMetadata.getterMetadata.signature.descriptor
         : generateGetterName(propertyMetadata.backingFieldSignature.memberName)
             + "()"
             + propertyMetadata.backingFieldSignature.descriptor;