+++ base/src/main/java/proguard/classfile/kotlin/visitor/KotlinEnumEntryVisitor.java
@@ -1,0 +1,12 @@
+package proguard.classfile.kotlin.visitor;
+
+import proguard.classfile.Clazz;
+import proguard.classfile.kotlin.KotlinClassKindMetadata;
+import proguard.classfile.kotlin.KotlinEnumEntryMetadata;
+
+public interface KotlinEnumEntryVisitor {
+  void visitAnyEnumEntry(
+      Clazz clazz,
+      KotlinClassKindMetadata kotlinClassKindMetadata,
+      KotlinEnumEntryMetadata kotlinEnumEntryMetadata);
+}